Publisher REST API pagination related changes
