+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/dto/APIListDTO.java
@@ -3,7 +3,7 @@ package org.wso2.carbon.apimgt.rest.api.publisher.v1.dto;
 import java.util.ArrayList;
 import java.util.List;
 import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.APIInfoDTO;
-import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.APIListPaginationDTO;
+import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.PaginationDTO;
 
 import io.swagger.annotations.*;
 import com.fasterxml.jackson.annotation.*;
@@ -22,16 +22,10 @@ public class APIListDTO  {
   private Integer count = null;
   
   
-  private String next = null;
-  
-  
-  private String previous = null;
-  
-  
   private List<APIInfoDTO> list = new ArrayList<APIInfoDTO>();
   
   
-  private APIListPaginationDTO pagination = null;
+  private PaginationDTO pagination = null;
 
   
   /**
@@ -44,32 +38,6 @@ public class APIListDTO  {
   }
   public void setCount(Integer count) {
     this.count = count;
-  }
-
-  
-  /**
-   * Link to the next subset of resources qualified.\nEmpty if no more resources are to be returned.\n
-   **/
-  @ApiModelProperty(value = "Link to the next subset of resources qualified.\nEmpty if no more resources are to be returned.\n")
-  @JsonProperty("next")
-  public String getNext() {
-    return next;
-  }
-  public void setNext(String next) {
-    this.next = next;
-  }
-
-  
-  /**
-   * Link to the previous subset of resources qualified.\nEmpty if current subset is the first subset returned.\n
-   **/
-  @ApiModelProperty(value = "Link to the previous subset of resources qualified.\nEmpty if current subset is the first subset returned.\n")
-  @JsonProperty("previous")
-  public String getPrevious() {
-    return previous;
-  }
-  public void setPrevious(String previous) {
-    this.previous = previous;
   }
 
   
@@ -89,10 +57,10 @@ public class APIListDTO  {
    **/
   @ApiModelProperty(value = "")
   @JsonProperty("pagination")
-  public APIListPaginationDTO getPagination() {
+  public PaginationDTO getPagination() {
     return pagination;
   }
-  public void setPagination(APIListPaginationDTO pagination) {
+  public void setPagination(PaginationDTO pagination) {
     this.pagination = pagination;
   }
 
@@ -104,8 +72,6 @@ public class APIListDTO  {
     sb.append("class APIListDTO {\n");
     
     sb.append("  count: ").append(count).append("\n");
-    sb.append("  next: ").append(next).append("\n");
-    sb.append("  previous: ").append(previous).append("\n");
     sb.append("  list: ").append(list).append("\n");
     sb.append("  pagination: ").append(pagination).append("\n");
     sb.append("}\n");
