+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/dto/LabelListDTO.java
@@ -3,6 +3,7 @@ package org.wso2.carbon.apimgt.rest.api.publisher.v1.dto;
 import java.util.ArrayList;
 import java.util.List;
 import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.LabelDTO;
+import org.wso2.carbon.apimgt.rest.api.publisher.v1.dto.PaginationDTO;
 
 import io.swagger.annotations.*;
 import com.fasterxml.jackson.annotation.*;
@@ -21,13 +22,10 @@ public class LabelListDTO  {
   private Integer count = null;
   
   
-  private String next = null;
-  
-  
-  private String previous = null;
-  
-  
   private List<LabelDTO> list = new ArrayList<LabelDTO>();
+  
+  
+  private PaginationDTO pagination = null;
 
   
   /**
@@ -44,32 +42,6 @@ public class LabelListDTO  {
 
   
   /**
-   * Link to the next subset of resources qualified.\nEmpty if no more resources are to be returned.\n
-   **/
-  @ApiModelProperty(value = "Link to the next subset of resources qualified.\nEmpty if no more resources are to be returned.\n")
-  @JsonProperty("next")
-  public String getNext() {
-    return next;
-  }
-  public void setNext(String next) {
-    this.next = next;
-  }
-
-  
-  /**
-   * Link to the previous subset of resources qualified.\nEmpty if current subset is the first subset returned.\n
-   **/
-  @ApiModelProperty(value = "Link to the previous subset of resources qualified.\nEmpty if current subset is the first subset returned.\n")
-  @JsonProperty("previous")
-  public String getPrevious() {
-    return previous;
-  }
-  public void setPrevious(String previous) {
-    this.previous = previous;
-  }
-
-  
-  /**
    **/
   @ApiModelProperty(value = "")
   @JsonProperty("list")
@@ -81,6 +53,18 @@ public class LabelListDTO  {
   }
 
   
+  /**
+   **/
+  @ApiModelProperty(value = "")
+  @JsonProperty("pagination")
+  public PaginationDTO getPagination() {
+    return pagination;
+  }
+  public void setPagination(PaginationDTO pagination) {
+    this.pagination = pagination;
+  }
+
+  
 
   @Override
   public String toString()  {
@@ -88,9 +72,8 @@ public class LabelListDTO  {
     sb.append("class LabelListDTO {\n");
     
     sb.append("  count: ").append(count).append("\n");
-    sb.append("  next: ").append(next).append("\n");
-    sb.append("  previous: ").append(previous).append("\n");
     sb.append("  list: ").append(list).append("\n");
+    sb.append("  pagination: ").append(pagination).append("\n");
     sb.append("}\n");
     return sb.toString();
   }
