+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/gen/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/dto/PaginationDTO.java
@@ -1,0 +1,109 @@
+package org.wso2.carbon.apimgt.rest.api.publisher.v1.dto;
+
+
+import io.swagger.annotations.*;
+import com.fasterxml.jackson.annotation.*;
+
+import javax.validation.constraints.NotNull;
+
+
+
+
+
+@ApiModel(description = "")
+public class PaginationDTO  {
+  
+  
+  
+  private Integer offset = null;
+  
+  
+  private Integer limit = null;
+  
+  
+  private Integer total = null;
+  
+  
+  private String next = null;
+  
+  
+  private String previous = null;
+
+  
+  /**
+   **/
+  @ApiModelProperty(value = "")
+  @JsonProperty("offset")
+  public Integer getOffset() {
+    return offset;
+  }
+  public void setOffset(Integer offset) {
+    this.offset = offset;
+  }
+
+  
+  /**
+   **/
+  @ApiModelProperty(value = "")
+  @JsonProperty("limit")
+  public Integer getLimit() {
+    return limit;
+  }
+  public void setLimit(Integer limit) {
+    this.limit = limit;
+  }
+
+  
+  /**
+   **/
+  @ApiModelProperty(value = "")
+  @JsonProperty("total")
+  public Integer getTotal() {
+    return total;
+  }
+  public void setTotal(Integer total) {
+    this.total = total;
+  }
+
+  
+  /**
+   * Link to the next subset of resources qualified.\nEmpty if no more resources are to be returned.\n
+   **/
+  @ApiModelProperty(value = "Link to the next subset of resources qualified.\nEmpty if no more resources are to be returned.\n")
+  @JsonProperty("next")
+  public String getNext() {
+    return next;
+  }
+  public void setNext(String next) {
+    this.next = next;
+  }
+
+  
+  /**
+   * Link to the previous subset of resources qualified.\nEmpty if current subset is the first subset returned.\n
+   **/
+  @ApiModelProperty(value = "Link to the previous subset of resources qualified.\nEmpty if current subset is the first subset returned.\n")
+  @JsonProperty("previous")
+  public String getPrevious() {
+    return previous;
+  }
+  public void setPrevious(String previous) {
+    this.previous = previous;
+  }
+
+  
+
+  @Override
+  public String toString()  {
+    StringBuilder sb = new StringBuilder();
+    sb.append("class PaginationDTO {\n");
+    
+    sb.append("  offset: ").append(offset).append("\n");
+    sb.append("  limit: ").append(limit).append("\n");
+    sb.append("  total: ").append(total).append("\n");
+    sb.append("  next: ").append(next).append("\n");
+    sb.append("  previous: ").append(previous).append("\n");
+    sb.append("}\n");
+    return sb.toString();
+  }
+}
