+++ components/apimgt/org.wso2.carbon.apimgt.rest.api.publisher.v1/src/main/java/org/wso2/carbon/apimgt/rest/api/publisher/v1/utils/mappings/SettingsMappingUtil.java
@@ -15,7 +15,7 @@
 * specific language governing permissions and limitations
 * under the License.
 */
-package org.wso2.carbon.apimgt.rest.api.publisher.v1.mappings;
+package org.wso2.carbon.apimgt.rest.api.publisher.v1.utils.mappings;
 
 
 import org.apache.commons.io.IOUtils;
